requirements.txt
torch
torchvision
matplotlib
Pillow
pandas 
mlflow
notebook
numpy
scikit-learn
matplotlib
python-box
pyYAML
tqdm
ensure
joblib
types-PyYAML
Flask
Flask-Cors
openpyxl
seaborn
tensorboard
uvicorn
python-multipart
torchmetrics
pycocotools
h5py
boto3